fix: nits
