+++ app/src/errors/components/ErrorBoundary.tsx
@@ -1,4 +1,3 @@
-import { toBifoldError } from '@/bcsc-theme/utils/error-utils'
 import { AppEventCode } from '@/events/appEventCode'
 import { Analytics } from '@/utils/analytics/analytics-singleton'
 import { AbstractBifoldLogger } from '@bifold/core'
@@ -7,6 +6,7 @@ import { useTranslation } from 'react-i18next'
 import { StyleSheet } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { AppError } from '../appError'
+import { toBifoldError } from '../errorHandler'
 import { ErrorInfoCard } from './ErrorInfoCard'
 
 interface ErrorBoundaryProps {
