+++ app/src/errors/errorHandler.test.ts
@@ -1,16 +1,14 @@
-import { extractErrorMessage, getErrorDefinition, getErrorDefinitionFromAppEventCode } from './errorHandler'
+import { UNKNOWN_APP_ERROR_STATUS_CODE } from '@/constants'
+import { AppEventCode } from '@/events/appEventCode'
+import { BifoldError } from '@bifold/core'
+import { AppError } from './appError'
+import {
+  extractErrorMessage,
+  getErrorDefinition,
+  getErrorDefinitionFromAppEventCode,
+  toBifoldError,
+} from './errorHandler'
 import { ErrorCategory, ErrorRegistry, ErrorSeverity } from './errorRegistry'
-
-// Mock dependencies
-jest.mock('react-native', () => ({
-  DeviceEventEmitter: {
-    emit: jest.fn(),
-  },
-  Platform: {
-    OS: 'ios',
-    select: jest.fn((obj) => obj.ios ?? obj.default),
-  },
-}))
 
 jest.mock('../../src/utils/analytics/analytics-singleton', () => ({
   Analytics: {
@@ -120,6 +118,54 @@ describe('errorHandler', () => {
       const definition = getErrorDefinitionFromAppEventCode('')
 
       expect(definition).toBeNull()
+    })
+  })
+
+  describe('toBifoldError', () => {
+    it('should convert a plain Error into a BifoldError with UNKNOWN_APP_ERROR_STATUS_CODE', () => {
+      const error = new Error('something broke')
+      error.stack = 'fake stack'
+
+      const result = toBifoldError('Title', 'Description', error)
+
+      expect(result).toBeInstanceOf(BifoldError)
+      expect(result.title).toBe('Title')
+      expect(result.description).toBe('Description')
+      expect(result.message).toBe('something broke')
+      expect(result.code).toBe(UNKNOWN_APP_ERROR_STATUS_CODE)
+      expect(result.stack).toBe('fake stack')
+    })
+
+    it('should convert an AppError into a BifoldError with its statusCode and fullMessage', () => {
+      const cause = new Error('technical details')
+      const appError = new AppError(
+        'App Error',
+        {
+          category: ErrorCategory.GENERAL,
+          appEvent: AppEventCode.GENERAL,
+          statusCode: 1000,
+        },
+        { cause, track: false }
+      )
+
+      const result = toBifoldError('Display Title', 'Display Description', appError)
+
+      expect(result).toBeInstanceOf(BifoldError)
+      expect(result.message).toBe(appError.fullMessage)
+      expect(result.title).toBe('Display Title')
+      expect(result.description).toBe('Display Description')
+      expect(result.code).toBe(1000)
+      expect(result.cause).toBe(cause)
+    })
+
+    it('should preserve the cause from the original error', () => {
+      const cause = new Error('root cause')
+      const error = new Error('wrapper')
+      error.cause = cause
+
+      const result = toBifoldError('T', 'D', error)
+
+      expect(result.cause).toBe(cause)
     })
   })
 })
