+++ app/src/screens/Splash.tsx
@@ -1,5 +1,5 @@
-import { toBifoldError } from '@/bcsc-theme/utils/error-utils'
 import { AppError, ErrorRegistry } from '@/errors'
+import { toBifoldError } from '@/errors/errorHandler'
 import { BCState } from '@/store'
 import {
   InfoBox,
